Change: Load fp16 weights on WebGPU so the model can actually be cached
fp32 was never cached. Cache Storage rejects put() for a body that size with an
opaque internal error long before the quota matters — measured on this project
as succeeding at 200 MB and failing at 256 MB, against a 5.6 GB quota — and
transformers.js only console.warns and carries on. The result was a 310 MB
download on every single page load, which is what "downloading model weights"
sitting there forever actually was for a returning visitor.

fp16 is 155 MB, so it fits under that ceiling, and it is much closer to fp32
than the q8 fallback. The WASM path keeps q8, which was already its dtype.

Verified against the production bundle served locally with the deployed
COOP/COEP pair, not the dev server:

  cold   155.7 MB downloaded, ready (WebGPU accelerated) in ~7s,
         model_fp16.onnx present in transformers-cache at 156 MB,
         badge flips to Local Cache
  warm   ready from cache in 547 ms
  wasm   forced device wasm still reaches init_complete with dtype q8 in ~2.9s

Quality against fp32 has not been compared by ear; fp16 is the conservative of
the two options for that reason.

+++ src/audio/kokoro-worker.js
@@ -109,9 +109,16 @@ async function handleInit(id, payload) {
 
   // WebGPU is several times faster than WASM and is what keeps lookahead ahead of
   // playback on long scripts. Fall back silently when it is missing or unusable.
+  //
+  // fp16 (155 MB) rather than fp32 (310 MB), because fp32 cannot be cached at
+  // all: Cache Storage rejects put() for bodies of that size with an opaque
+  // internal error long before the quota is reached — measured on this project
+  // as succeeding at 200 MB and failing at 256 MB — so transformers.js warns and
+  // moves on, and every visit re-downloads the whole model. fp16 fits under that
+  // ceiling and is far closer to fp32 than the q8 fallback is.
   const attempts = [];
   if (payload.device !== 'wasm' && typeof navigator !== 'undefined' && navigator.gpu) {
-    attempts.push({ device: 'webgpu', dtype: 'fp32' });
+    attempts.push({ device: 'webgpu', dtype: 'fp16' });
   }
   attempts.push({ device: 'wasm', dtype: 'q8' });
